+++ pkg/analyzer_plugin/lib/plugin/plugin.dart
@@ -198,14 +198,19 @@ abstract class ServerPlugin {
   /// one or more files. The implementation may check if these files should
   /// be analyzed, do such analysis, and send diagnostics.
   ///
-  /// By default invokes [analyzeFiles].
+  /// By default invokes [analyzeFiles] only for files that are analyzed in
+  /// this [analysisContext].
   Future<void> handleAffectedFiles({
     required AnalysisContext analysisContext,
     required List<String> paths,
   }) async {
+    final analyzedPaths = paths
+        .where(analysisContext.contextRoot.isAnalyzed)
+        .toList(growable: false);
+
     await analyzeFiles(
       analysisContext: analysisContext,
-      paths: paths,
+      paths: analyzedPaths,
     );
   }