+++ pkg/analyzer_plugin/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.11.1
+- Call `analyzeFiles` from `handleAffectedFiles` only for files that are
+  analyzed in this analysis context.
+
 ## 0.11.0
 - Using `AnalysisContextCollection` and `AnalysisContext` for analysis.